fix: Prevent uncontrolled to controlled input error by providing default empty string value
Error
```
src/components/ui/input.tsx (7:3) @ Input

   5 | function Input({ className, type, ...props }: React.ComponentProps<"input">) {
   6 | 	return (
>  7 | 		<input
     | 		^

A component is changing an uncontrolled input to be controlled. This is likely caused by the value changing from undefined to a defined value, which should not happen. Decide between using a controlled or uncontrolled input element for the lifetime of the component. More info: https://react.dev/link/controlled-components
```